Move the owned object index out of filesystemstore and into its own module
